improve error message for embedding mismatch
